fix: move providers types file
