+++ src/providers/slack.ts
@@ -1,7 +1,7 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
 
 import { OAuth2Client, OAuth2ClientConfig } from "../../deps.ts";
-import type { WithScope } from "./_types.ts";
+import type { WithScope } from "./types.ts";
 
 /**
  * Creates an OAuth 2.0 client with Slack as the provider.
